2016.1.15
购物车，订单，结算，新农代表，我的订单

+++ xinnongreniOS/Classes/Mine/Controller/XNRMakeSureOrderInfo_VC.m
@@ -247,7 +247,7 @@
             }
             
         }else{
-            totalPrice = totalPrice + model.goodsCount.intValue * model.unitPrice;
+            totalPrice = totalPrice + model.goodsCount.intValue * model.unitPrice.floatValue;
         }
     }
